feat: 对战若干优化
- 咖啡豆夜间不可选取
- 平衡调整:夜间蘑菇花费+25
- 开局预放置的阳光菇变为大号
- 刷牌模式:咖啡豆相关逻辑优化

+++ app/src/main/cpp/PvZ/src/Lawn/Widget/VSSetupMenu.cpp
@@ -629,10 +629,10 @@ void VSSetupMenu::OnStateEnter(VSSetupState theState) {
         if (tcpClientSocket >= 0) {
             B1x8_Event event = {
                 {EventType::EVENT_SERVER_VSSETUP_ADDON_BUTTON_INIT},
-                uint8_t(gVSSetupAddonWidget->mExtraPacketsMode),
-                uint8_t(gVSSetupAddonWidget->mExtraSeedsMode),
-                uint8_t(gVSSetupAddonWidget->mBanMode),
-                uint8_t(gVSSetupAddonWidget->mBalancePatchMode),
+                gVSSetupAddonWidget->mExtraPacketsMode,
+                gVSSetupAddonWidget->mExtraSeedsMode,
+                gVSSetupAddonWidget->mBanMode,
+                gVSSetupAddonWidget->mBalancePatchMode,
             };
             sendWithSize(tcpClientSocket, &event, sizeof(B1x8_Event), 0);
         }
